requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
requests
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
elevenlabs
